fix: bugfix for cpu sampler
When we use the proc_cpuinfo sampler on systems that do not report
a per processor frequency, the sampler will panic with a divide by
zero.

This change fixes the logic so that we avoid a divide by zero when
/proc/cpuinfo doesn't parse as we expect. It also changes the logic
on how we count the number of online cores to get a count even when
the system doesn't report core frequencies.