fix: set packet latency update mode to no_exist
Changes the packet latency sampler behavior to update tracking
entries when a packet is received only if there is not an existing
timestamp entry for that socket.

Removes unnecessary filtering variables and logic.

This means that if multiple packets are received before userspace
processes them, that the longest latency will be counted instead
of the shortest.