fix: histogram mmap length check
The length check of the mmap'd histogram buckets used the exact
number of buckets in the histogram. This number needs to be rounded
up to match the next full page size.

Instead of using the `HISTOGRAM_BUCKETS` const, we use the
`HISTOGRAM_PAGES` and `PAGE_SIZE` to validate the length of the
mmap'd buckets.

This results in the fast path being used.