bpf optimization syscall sampler (#195)
Optimize the syscall bpf sampler by replacing the hashmap with an
array for all possible PIDs. This reduces the amount of overhead
per event at the expense of more memorty utilization.

This also replaces the chained if-else block with a nested set in
order to reduce the number of branches per update.