fix: report memory usage gauges in bytes
Currently, the metadata units and actual measurement units for the
memory usage gauges (`memory/total`, `memory/free`, ...) do not
agree.

This change updates the sampler so that the gauges are properly
reported in bytes.

While this is more disruptive than changing the unit metadata, this
approach leads to better consistency across our units of
measurement.