refactor bpf histogram indexing
The previous BPF histogram indexing was based on a reduction of the
algorithm in an older version of the histogram crate that had both
an a and b component. The latest histogram crate replaced those
with a single grouping power.

This change brings the indexing more in-line with the Rust version.
Specific reductions and optimizations are made based on the specific
values that we use throughout Rezolus.

We also add indexing for grouping powers 4, 5, and 6 so that in the
future we can use lower-resolution histograms for cases where we are
doing per-group histograms (eg: per group of PIDs/TIDs) and would
otherwise be adding a fair amount of memory footprint to Rezolus to
enable scoped histograms.