refactor bpf histogram indexing (#249)
The previous BPF histogram indexing was based on a simplification of the
algorithm in an older version of the histogram crate that had both
an a and b component. The latest histogram crate replaced those
with a single grouping power.

This change brings the indexing more in-line with the Rust version.
Specific simplifications and optimizations are made based on the
parameters that we use throughout Rezolus.