add cpu busy time (#256)
Adds metrics for cpu busy time to make cpu utilization easier to
calculate. This metric is not exported in the prometheus exposition
because all the non-idle states can be trivially summed. However,
for users of other metrics systems or consumers of the raw msgpack
output, it can be more convenient to work with just busy and idle
times and not worry about enumerating and summing the various other
states.