fix configuration logic (#273)
Previously, when a sampler config section is partially defined, it inherits all defaults in the code as opposed to the defaults provided in the config. This means that the configurations may sometimes need to be overly verbose.

This change addresses the logic in the config module such that all sampler config fields are optional and inherit from the config defaults and then the coded defaults. This brings the behavior in-line with expectations.

Also fixes the cpu perf sampler initialization to allow running without BPF perf events being enabled.