deps: update histogram, metriken, and ringlog
Update dependencies to move towards a single version
of histogram (0.11.0) and metriken (0.7.0). This is
a breaking change in the interface of calls to percentile
functions, which can now return null when the histogram
is empty.

The prometheus_stats and human_stats functions are the
consumers of the percentiles function and when it returns
a null, they do not expose any of the percentile
metrics.