add configurable snapshot interval (#286)
Previously the snapshot interval was fixed to be on a secondly
basis. For those who scrape Rezolus at coarser intervals, it would
be more appropriate to align the snapshot interval with the scrape
interval so that the percentiles cover a window spanning the entire
period between consecutive scrapes of the metrics endpoint.