split syscall sampler counts from latency
Splits the syscall bpf sampler into two parts. One that tracks just
the syscall counts and is expected to be lighter weight, and pares
down the original latency sampler to just handle the latency
metrics.

The intent is that for some sensitive workloads, we may be able to
have syscall count tracking enabled even with the latency metrics
disabled.