fix gaps in msgpack exposition (#408)
With the move to custom metric types for grouped metrics, we have
gaps in the msgpack exposition.

This PR closes these gaps by using custom snapshot logic to produce
a metriken-exposition compatible snapshot. The key difference is
that we now use opaque numeric IDs for the column names and put the
base metric name in the `metric` field in the column metadata. This
allows us to have duplicate metric names which will occur when
we are using attributes to facet along cpu/cgroup or other
meaningful properties (read/write, receive/transmit, ...)