fix: remove slash rewriting in prometheus exposition (#444)
The rewriting of `/` to `_` in prometheus exposition is no longer
needed and complicates the exposition of the cgroup names in a
legible manner.

This removes the string replacement that was being done. All metric
names have already had the slashes replaced, so this change does
not break the exposition format or metric names.

After this change, cgroup names as collected on the prometheus
endpoint will have forward slashes used to indicate the cgroup
hierarchy.