fix: flight-recorder off-by-one error while writing artifact (#447)
Fixes a simple off-by-one error when writing out the packed msgpack
artifact.

Previously, the oldest sample mistakenly is placed at the end of
the artifact. This may cause issues if the consumer expects the
timestamps to be monotonically nondecreasing.

Corrects the error and fixes up the comments so it's clearer going
forward.