report time as start of snapshot
Previously, we used the end of the refresh of all samplers as the
timestamp for our metrics snapshots. With the changes in metriken
to support reporting both a timestamp and duration in the snapshot,
we update to using the start edge and the new snapshot v2 format.

Prometheus exposition is updated to also use the start edge as this
time is likely to better reflect when metrics are read out.