add summarize subcommand
Adds a new subcommand 'summarize' which serves as a proxy which
scrapes the msgpack endpoint periodically and produces summary
metrics (percentiles) from the histograms.

This PR also adds the `rezolus-summarize` service definition which
by default will connect to the default Rezolus port (4242) on
localhost and listen on port 4243 on all addresses. This service
is disabled by default. This is configured for secondly scraping
and the interval should be set to match the scrape interval of
metrics collection.

In environments where collection of percentiles instead of full
histograms is desired, the `rezolus-summarize` service should be
enabled, and the metrics collection should be pointed at port 4243
instead of the Rezolus agent's default port 4242.