add rezolus exporter subcommand
Adds an exporter which acts as a client of the Rezolus agent and
provides a prometheus compatible endpoint.

To be able to create summary metrics from histograms, periodic
snapshots must be taken to produce delta histograms from the
otherwise free-running histograms. To keep the Rezolus agent as a
passive component, the exposition is moved into a separate
subcommand so that periodic snapshotting can be introduced.

This change introduces the new exporter subcommand, which by
default will listen on port :4241 and connect to a Rezolus agent
on localhost port :4242.