use async http client (#505)
Previously, we were using the blocking Reqwest client in an async
context. While done intentionally, this does result in a panic in
debug builds.

Changes the recorder, exporter, and hindsight tools to use an async
client in the async context.

Updates the ci smoketest to be a combination of the agent and
recorder.