Viewer: heatmap perf improvement 2
Not a documented param for a custom series (as it is for a 'heatmap' series)
but it still works. It controls how much rendering work it's willing to do at once (in one frame). Bigger is better both for the obvious reason and because each render update requires reevaluating its past work.

Setting this too high causes page-wide hitches during the time it's drawing the heatmap,
so it's somewhat of a tradeoff.