Viewer refactor - expand purview of type-specific echart config fns
I'm not 100% happy with this, but I think we need these (nearly-)pure
functions to have more side effects, so I'm making it explicit.

Previously the official scope of them was to return an echarts configuration
("option"); now they responsible for setting the option on the echarts
object directly as well as for potentially other imperative things, like
setting up caching of processed data and adding type-specific event handlers.