Viewer: cache and preload section data for better responsiveness
In sample data, some sections take multiple seconds for the request.
In this case, compared to many modern websites it is not that much
memory to retain them all, and it is even valuable to preload them as
a multisecond delay feels slow.

In cases with significantly more data, there might be a scenario where we
regret this and would like to handle it with more prudence, but the scale
at which we arguably can't afford the memory is a very narrow target.
(Before too long, we'd have too much data in even a single section to handle
gracefully.)