feature: use pinned threads to collect cpu_frequency samples
Previously, the perf groups for cpu_frequency sampler were read
serially and blocked the sampler. On some systems, it can take
several milliseconds to collect the readings in this way (m7i.8xl).

This change uses the same execution model we use in the BPF based
perf samplers, where we have one thread per core and use a sync
primitive (condvar w/ notify) to make the operation async from the
caller's perspective. To refresh the sampler, we trigger collection
on all perf threads and do a joined await.

As a result the cpu_frequency sampler latency should improve on all
systems as the collection will happen in parallel. The benefits are
more noticeable in virtualized environments.