fix: viewer page not found error in production build
When the Rezolus Viewer is run from a production build, we expect
that all the served assets will be from copies embedded in the
binary. This allows the Viewer to be run from anywhere.

Unfortunately, in 5.0.0 release, the index.html page was not served
from an embedded copy. This results in a 404 unless the Viewer is
run from a cloned copy of the repo.

This change fixes addresses this by serving the index from the
embedded copy unless developer-mode is enabled at build time.