feat(agent): enable realtime scheduling on linux
Adds scheduler configuration to the Rezolus Agent. This allows for
a scheduler policy to be set and to adjust the priority/niceness.

By default, Rezolus Agent will run as a realtime service with a
priority of 15 to balance between realtime responsiveness and
safety.

As a result, metrics snapshots will be produced with better timing
guarantees.