fix(agent): fix cpu accounting bug when PID is re-used (#571)
Fixes a CPU usage accounting bug that occurs when a PID is re-used.
This results in a large increment to the core/cgroup counters and
appears as non-monotonic behavior for the counter.

Detects unreasonably large increments and treats them as counter
resets instead. This allows for proper accounting on both wrap and
reset cases.

* undo clang format