fix(agent): capture and log messages from libbpf
Previously, libbpf would log messages to stderr if there were any
failures in creating or attaching probes. With the introduction of
driver-specific probes in #565 there are now probes which we expect
will fail when those drivers are not present in the running kernel.

These error messages, though harmless, may mislead users and cause
confusion.

This change captures the libbpf messages and logs them at debug
level. Further, we demote the attach failure message from warn to
debug for this specific case where the symbols are not found. This
results in cleaner output from the agent while allowing access to
the messages at debug level when necessary.