feat(viewer): convert viewer to promql (#604)
This is a large change to the viewer which adds a promql endpoint
and moves dashboards from using pre-generated JSON using Rust to
specify the timeseries transformations to using promql to express
the timeseries for each chart.

This also adds a new query explorer page which allows specifying
promql to plot.

The cgroup view has been reworked to use a two column view of
other vs individual cgroups.