feat(mcp): add correlation analysis tool (#617)
Adds a tool to the MCP server for doing correlation analysis. This performs cross-correlation between two queries. If a query returns multiple series, the analysis is performed on the cartesian product so that correlations between individual series can be examined. For example, seeing if a system metric is correlated with per-cgroup metrics.

Adds CLI invocation for the describe recording and analyze correlation tools so they can be run without using a MCP client.