add http endpoint to hindsight
Adds an http endpoint to hindsight so that a user can remotely
trigger a capture to disk, or have a capture streamed back to them.

This provides an interface for capturing a specific time range,
which is useful when the interesting portion of the buffer is
significantly shorter than the total buffer.